set params outside of caching
